Fixed eslint warning
