+++ src/contexts/ThemeContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+import { ThemeMode } from '../types/ThemeMode';
+
+interface ThemeContextType {
+  theme: ThemeMode;
+  setTheme: (theme: ThemeMode) => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
