+++ src/contexts/ThemeContext.tsx
@@ -1,6 +1,4 @@
 import React, {
-  createContext,
-  useContext,
   useState,
   useEffect,
   ReactNode,
@@ -11,21 +9,7 @@ import {
 } from '@mui/material/styles';
 import { getOptionsFromStorage } from '../functions/storage';
 import { ThemeMode } from '../types/ThemeMode';
-
-interface ThemeContextType {
-  theme: ThemeMode;
-  setTheme: (theme: ThemeMode) => void;
-}
-
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  if (!context) {
-    throw new Error('useTheme must be used within a ThemeProvider');
-  }
-  return context;
-};
+import { ThemeContext } from './ThemeContext';
 
 interface ThemeProviderProps {
   children: ReactNode;
@@ -83,7 +67,7 @@ export const ThemeProvider: React.FC<ThemeProviderProps> = ({ children }) => {
     },
   });
 
-  const value: ThemeContextType = {
+  const value = {
     theme,
     setTheme,
   };
