+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useDialog } from './useDialog';
 export { useSnackbar } from './useSnackbar';
+export { useTheme } from './useTheme';
